# powerapps/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SafeTrace Local API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/batches/analyze": {
      "post": {
        "summary": "Analyze Batch",
        "operationId": "analyze_batch_api_batches_analyze_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_batch_api_batches_analyze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "summary": "Job Status",
        "operationId": "job_status_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Job",
        "operationId": "delete_job_api_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/result": {
      "get": {
        "summary": "Job Result",
        "operationId": "job_result_api_jobs__job_id__result_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/batches/{batch_id}": {
      "get": {
        "summary": "Batch Status",
        "operationId": "batch_status_api_batches__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Batch",
        "operationId": "delete_batch_api_batches__batch_id__delete",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reports/{job_id}/technical-json": {
      "get": {
        "summary": "Technical Json",
        "operationId": "technical_json_api_reports__job_id__technical_json_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchRejectedFile": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "sourceRelativePath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcerelativepath"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "reason"
        ],
        "title": "BatchRejectedFile"
      },
      "BatchResponse": {
        "properties": {
          "batchId": {
            "type": "string",
            "title": "Batchid"
          },
          "status": {
            "type": "string",
            "enum": [
              "importing",
              "queued",
              "waiting_for_capacity",
              "running",
              "retry_wait",
              "paused",
              "recovering",
              "completed",
              "completed_with_failures",
              "failed",
              "partial",
              "cancelled"
            ],
            "title": "Status"
          },
          "sourceFilename": {
            "type": "string",
            "title": "Sourcefilename"
          },
          "batchDisplayLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batchdisplaylabel"
          },
          "importKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importkey"
          },
          "acceptedFiles": {
            "items": {
              "$ref": "#/components/schemas/BatchAcceptedFile"
            },
            "type": "array",
            "title": "Acceptedfiles"
          },
          "rejectedFiles": {
            "items": {
              "$ref": "#/components/schemas/BatchRejectedFile"
            },
            "type": "array",
            "title": "Rejectedfiles"
          },
          "jobIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Jobids"
          },
          "statusCounts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Statuscounts"
          },
          "groupSummaries": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Groupsummaries"
          },
          "hierarchy": {
            "type": "object",
            "title": "Hierarchy"
          },
          "throughput": {
            "type": "object",
            "title": "Throughput"
          },
          "paused": {
            "type": "boolean",
            "title": "Paused",
            "default": false
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          },
          "persistenceWarning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persistencewarning"
          }
        },
        "type": "object",
        "required": [
          "batchId",
          "status",
          "sourceFilename",
          "acceptedFiles",
          "rejectedFiles",
          "jobIds",
          "statusCounts",
          "createdAt",
          "updatedAt"
        ],
        "title": "BatchResponse"
      },
...
      "Body_analyze_batch_api_batches_analyze_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "fps": {
            "type": "number",
            "title": "Fps",
            "default": 1.0
          },
          "topK": {
            "type": "integer",
            "title": "Topk",
            "default": 5
          },
          "enableVlm": {
            "type": "boolean",
            "title": "Enablevlm",
            "default": false
          },
          "vlmProfile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlmprofile"
          },
          "vlmEnabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlmenabled"
          },
          "useCaseProfile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usecaseprofile"
          },
          "device": {
            "type": "string",
            "enum": [
              "auto",
              "cpu",
              "cuda"
            ],
            "title": "Device",
            "default": "auto"
          },
          "reviewMode": {
            "type": "string",
            "title": "Reviewmode",
            "default": "fast_local"
          },
          "relativePaths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relativepaths"
          },
          "importKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importkey"
          }
        },
        "type": "object",
        "required": [
          "files",
          "query"
        ],
        "title": "Body_analyze_batch_api_batches_analyze_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStatusResponse": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "waiting_for_capacity",
              "running",
              "retry_wait",
              "recovering",
              "paused",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "progressPercent": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progresspercent"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "currentStep": {
            "type": "string",
            "title": "Currentstep"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "metrics": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "componentDiagnostics": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Componentdiagnostics"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "queuedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queuedat"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "failedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failedat"
          },
          "cancelledAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelledat"
          },
          "elapsedSeconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsedseconds"
          },
          "queueWaitSeconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queuewaitseconds"
          },
          "analysisRuntimeSeconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysisruntimeseconds"
          },
          "heartbeatAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeatat"
          },
          "persistenceWarning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persistencewarning"
          },
          "manifestPersistenceWarning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifestpersistencewarning"
          },
          "requestedModeLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestedmodelabel"
          },
          "requestedVisualExplanationMode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestedvisualexplanationmode"
          },
          "actualExplanationMode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actualexplanationmode"
          },
          "finalExplanationSource": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalexplanationsource"
          },
          "explanationOutcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanationoutcome"
          },
          "explanationOutcomeLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanationoutcomelabel"
          },
          "vlmAttempted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlmattempted"
          },
          "lightweightVlmAttempted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lightweightvlmattempted"
          },
          "enhancedVlmAttempted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enhancedvlmattempted"
          },
          "vlmAccepted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlmaccepted"
          },
          "lightweightVlmAccepted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lightweightvlmaccepted"
          },
          "enhancedVlmAccepted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enhancedvlmaccepted"
          },
          "vlmFallbackReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlmfallbackreason"
          },
          "vlmFallbackReasonLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlmfallbackreasonlabel"
          },
          "actualDeviceLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actualdevicelabel"
          },
          "jobRuntimeLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jobruntimelabel"
          },
          "engineRuntimeSummary": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engineruntimesummary"
          },
          "sourceMetadata": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcemetadata"
          },
          "retryCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrycount"
          },
          "maxAttempts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxattempts"
          },
          "nextRetryAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextretryat"
          },
          "recoveryAction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recoveryaction"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "status",
          "progress",
          "progressPercent",
          "stage",
          "message",
          "currentStep"
        ],
        "title": "JobStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}